rename directory to work with naming convention for CI
